feat: Реализовать верстку экрана ввода жидкости

+++ featureLiquid/api/src/main/java/ru/health/featureliquid/api/domain/model/BottleType.kt
@@ -1,7 +1,7 @@
 package ru.health.featureliquid.api.domain.model
 
-enum class BottleType(val volume: Int) {
-    SMALL(30),
-    TALL(60),
-    LARGE(120)
+enum class BottleType(val volume: Float) {
+    SMALL(30f),
+    TALL(60f),
+    LARGE(120f)
 }
